requirements.txt
flask==3.0.0
python-telegram-bot==20.7
apscheduler==3.10.4
gunicorn==21.2.0